% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamlahti.R
\docType{data}
\name{Lee2023}
\alias{Lee2023}
\title{Metagenome assembly data}
\format{
A TreeSummarizedExperiment with 4182 features and 5195 samples. The
rowData contains taxonomic information at Kingdom, Phylum, Class, Order,
Family and Genus level. See \href{https://ncbi-codeathons.github.io/amr-2024-team-lahti/articles/data_summary.html}{dataset summary}.
}
\usage{
data(Lee2023)
}
\description{
Lee2023 includes a total of 9,251 human microbiome samples from multiple body
sites (7718 stool, 783 oral cavity, 410 skin, 150 airway, 93 nasal cavity,
88 vagina, and 9 milk) which had been sequenced in various studies.
}
\references{
Lee, K., Raguideau, S., Sirén, K. et al.
Population-level impacts of antibiotic usage on the human gut microbiome.
Nat Commun 14, 1191 (2023). \url{https://doi.org/10.1038/s41467-023-36633-7}
}
\seealso{
\code{\link{teamlahti-datasets}}
}
\author{
Lee, K., Raguideau, S., Sirén, K. et al.
}
\keyword{datasets}
