% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamlahti.R
\docType{data}
\name{teamlahti-datasets}
\alias{teamlahti-datasets}
\title{teamlahti datasets}
\description{
teamlahti provides Lee2023 dataset derived from independent experimental
studies.The dataset represents instances of the TreeSummarizedExperiment to
serve as tools to practice the teamlahti functionality.
}
\details{
Currently, the following datasets are available:
\itemize{
\item{\code{\link{Lee2023}}: A TreeSummarizedExperiment with 9,251 samples}
}
}
\examples{
# Load dataset from teamlahti
library(teamlahti)
data("Lee2023", package = "teamlahti")

# In this case, the dataset is a TreeSE, so it is renamed as tse
tse <- Lee2023

# Print summary
tse
}
\keyword{datasets}
