% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamlahti.R
\docType{package}
\name{teamlahti-package}
\alias{teamlahti}
\alias{teamlahti-package}
\title{\code{teamlahti} Package.}
\description{
\code{teamlahti} aims to develop \href{https://www.bioconductor.org/}{R/Bioconductor}
tools to analyze AMR data by integrating publicly available human gut
metagenomes with ARG profiles (e.g.
\href{https://www.nature.com/subjects/metagenomics}{metagenomics}) and
epidemiological, transmission data to study AMR gene co-migration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NCBI-Codeathons/amr-2024-team-lahti}
  \item Report bugs at \url{https://github.com/NCBI-Codeathons/amr-2024-team-lahti/issues}
}

}
\author{
\strong{Maintainer}: Muluh Geraldson \email{rysdaena8@gmail.com} (\href{https://orcid.org/0000-0002-9721-5192}{ORCID})

Authors:
\itemize{
  \item Leo Lahti \email{leo.lahti@iki.fi} (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
  \item Akewak Jeba \email{akjeba@utu.fi} (\href{https://orcid.org/0009-0007-1347-7552}{ORCID})
  \item Nitin Bayal \email{nitin.bayal@utu.fi} (\href{https://orcid.org/0000-0002-4840-9898}{ORCID})
  \item Shivang Bhanushali \email{shivang.bhanushali@utu.fi} (\href{https://orcid.org/0009-0007-8277-3662}{ORCID})
  \item Dattaray Mongad \email{datta.mongad@gmail.com} (\href{https://orcid.org/0000-0003-4801-7410}{ORCID})
  \item Mahkameh Salehi \email{mahksa@utu.fi} (\href{https://orcid.org/0009-0009-7153-226X}{ORCID})
}

}
