% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamlahti.R
\docType{data}
\name{Lee2023}
\alias{Lee2023}
\title{Metagenome assembly data}
\format{
A TreeSummarizedExperiment with 4182 features and 5195 samples. The
rowData contains taxonomic information at Kingdom, Phylum, Class, Order,
Family and Genus level. The colData includes:

\describe{
\item{Tier_1.Exclusion_before_analysis}
\item{Tier_2.Recover_all_40_SCGs}
\item{Tier_3.Adult_stool_from_well_sampled_countries}
\item{Tier_4.Adult_stool_from_healthy_subjects_not_currently_on_antibiotic}
\item{Study}
\item{Sample_ID}
\item{Sample_assembly_ID}
\item{NumReads}
\item{NumBases}
\item{MedianReadLength}
\item{BodySite}
\item{BodySubsite}
\item{AgeYears}
\item{AgeCategory}
\item{InfantAge}
\item{Westernized}
\item{Country}
\item{Gender}
\item{BMI}
\item{Disease}
\item{antibiotic_exposure_status_descriptive}
\item{antibiotic_current_use_binary}
}
}
\usage{
data(Lee2023)
}
\description{
Lee2023 includes a total of 9,251 human microbiome samples from multiple body
sites (7718 stool, 783 oral cavity, 410 skin, 150 airway, 93 nasal cavity,
88 vagina, and 9 milk) which had been sequenced in various studies.
}
\references{
Lee, K., Raguideau, S., Sirén, K. et al.
Population-level impacts of antibiotic usage on the human gut microbiome.
Nat Commun 14, 1191 (2023). \url{https://doi.org/10.1038/s41467-023-36633-7}
}
\seealso{
\code{\link{teamlahti-datasets}}
}
\author{
Lee, K., Raguideau, S., Sirén, K. et al.
}
\keyword{datasets}
